add multiple projects
