+++ kotlin/build.gradle.kts
@@ -1,45 +1,6 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    kotlin("jvm") version "1.7.20"
-    `maven-publish`
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.13.4")
-    testImplementation(kotlin("test"))
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-
-tasks.withType<KotlinCompile> {
-    kotlinOptions.jvmTarget = "17"
-}
-
-publishing {
-    repositories {
-        maven {
-            name = "GitHubPackages"
-            url = uri("https://maven.pkg.github.com/warnermedia/rules-engine")
-            credentials {
-                username = System.getenv("GITHUB_ACTOR")
-                password = System.getenv("GITHUB_TOKEN")
-            }
-        }
-    }
-    publications {
-        create<MavenPublication>("maven") {
-            groupId = "com.warnermedia"
-            artifactId = "rulesengine"
-            version = "1.8.1"
-
-            from(components["java"])
-        }
-    }
-}
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This is a general purpose Gradle build.
+ * Learn more about Gradle by exploring our samples at https://docs.gradle.org/7.4.2/samples
+ */
