+++ kotlin/core/src/main/kotlin/core/ConditionEvaluationOptions.kt
@@ -1,0 +1,9 @@
+package core
+
+/**
+ * Class defining options passed during evaluation of a condition
+ */
+class ConditionEvaluationOptions(
+    val upcastFactValues: Boolean,
+    val undefinedFactEvaluationType: UndefinedFactEvaluation
+)
