+++ kotlin/core/src/main/kotlin/core/ConditionJoiner.kt
@@ -1,0 +1,8 @@
+package core
+
+/**
+ * Enum defining possible ways of joining multiple conditions within a rule
+ */
+enum class ConditionJoiner {
+    AND, OR
+}
