+++ kotlin/core/src/main/kotlin/core/EngineEvaluationType.kt
@@ -1,0 +1,8 @@
+package core
+
+/**
+ * Enum defining different ways an engine will evaluate the rule set
+ */
+enum class EngineEvaluationType {
+    ALL, FIRST_ERROR, FIRST_FAILURE, FIRST_SUCCESS
+}
