+++ kotlin/core/src/main/kotlin/core/EvaluationResult.kt
@@ -1,0 +1,10 @@
+package core
+
+/**
+ * Class defining result of an engine evaluation of the rule set
+ */
+class EvaluationResult(val ruleEvaluations: ArrayList<RuleResult>, val exitCriteria: ExitCriteria) {
+    override fun toString(): String {
+        return ruleEvaluations.toString()
+    }
+}
