+++ kotlin/core/src/main/kotlin/core/Operator.kt
@@ -1,0 +1,6 @@
+package core
+
+/**
+ * Class defining an operation type along with the value to perform it against
+ */
+class Operator(val operatorType: OperatorType, val operatorValue: Any)
