+++ kotlin/core/src/main/kotlin/core/RuleEvaluationOptions.kt
@@ -1,0 +1,10 @@
+package core
+
+/**
+ * Class defining options passed to a Rule during runtime evaluation
+ */
+class RuleEvaluationOptions(
+    val upcastFactValues: Boolean,
+    val undefinedFactEvaluationType: UndefinedFactEvaluation,
+    val storeRuleEvaluationResults: Boolean
+)
