+++ kotlin/core/src/main/kotlin/core/UndefinedFactEvaluation.kt
@@ -1,0 +1,8 @@
+package core
+
+/**
+ * Enum defining how to evaluate a runtime fact that is not defined
+ */
+enum class UndefinedFactEvaluation {
+    EVALUATE_TO_TRUE, EVALUATE_TO_FALSE, EVALUATE_TO_SKIPPED
+}
