+++ kotlin/gradle.properties
@@ -1,1 +1,3 @@
 kotlin.code.style=official
+packageGroupId=com.warnermedia.rulesengine
+packageVersionId=0.0.1
