--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/ConditionEvaluationOptions.kt
@@ -1,9 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Class defining options passed during evaluation of a condition
- */
-class ConditionEvaluationOptions(
-    val upcastFactValues: Boolean,
-    val undefinedFactEvaluationType: UndefinedFactEvaluation
-)
