--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/ConditionJoiner.kt
@@ -1,8 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Enum defining possible ways of joining multiple conditions within a rule
- */
-enum class ConditionJoiner {
-    AND, OR
-}
