--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/EngineEvaluationType.kt
@@ -1,8 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Enum defining different ways an engine will evaluate the rule set
- */
-enum class EngineEvaluationType {
-    ALL, FIRST_ERROR, FIRST_FAILURE, FIRST_SUCCESS
-}
