--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/EvaluationResult.kt
@@ -1,10 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Class defining result of an engine evaluation of the rule set
- */
-class EvaluationResult(val ruleEvaluations: ArrayList<RuleResult>, val exitCriteria: ExitCriteria) {
-    override fun toString(): String {
-        return ruleEvaluations.toString()
-    }
-}
