--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/ExitCriteria.kt
@@ -1,18 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Class defining possible exit criteria for an engine evaluation
- */
-sealed class ExitCriteria {
-    class EarlyExit(val ruleResult: RuleResult) : ExitCriteria()
-
-    object NormalExit : ExitCriteria()
-
-    fun isEarlyExit(): Boolean {
-        return this is EarlyExit
-    }
-
-    fun isNormalExit(): Boolean {
-        return this is NormalExit
-    }
-}
