--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/Operator.kt
@@ -1,6 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Class defining an operation type along with the value to perform it against
- */
-class Operator(val operatorType: OperatorType, val operatorValue: Any)
