--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/RuleEvaluationOptions.kt
@@ -1,10 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Class defining options passed to a Rule during runtime evaluation
- */
-class RuleEvaluationOptions(
-    val upcastFactValues: Boolean,
-    val undefinedFactEvaluationType: UndefinedFactEvaluation,
-    val storeRuleEvaluationResults: Boolean
-)
