--- kotlin/src/main/kotlin/com/warnermedia/rulesengine/UndefinedFactEvaluation.kt
@@ -1,8 +1,0 @@
-package com.warnermedia.rulesengine
-
-/**
- * Enum defining how to evaluate a runtime fact that is not defined
- */
-enum class UndefinedFactEvaluation {
-    EVALUATE_TO_TRUE, EVALUATE_TO_FALSE, EVALUATE_TO_SKIPPED
-}
